Rename Study_network.py to Study_Models.py for improved clarity and consistency
Removed Vectorizer.py as it is now not used
